requirements.txt
flask==2.3.2
gunicorn==21.2.0
python-docx==1.1.0
python-pptx==0.6.23
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.1.0
lxml==4.9.3

